+++ demisto_sdk/commands/common/hook_validations/generic_field.py
@@ -68,7 +68,7 @@ class GenericFieldValidator(ContentEntityValidator):
         indicator_field_unsearchable = self.current_file.get('unsearchable', True)
         if indicator_field_unsearchable:
             return True
-        error_message, error_code = Errors.unsearchable_key_should_be_true_gf()
+        error_message, error_code = Errors.unsearchable_key_should_be_true_generic_field()
         if self.handle_error(error_message, error_code, file_path=self.file_path):
             return False
         return True