+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 * Fixed an issue where the **update-release-notes** command crashes with `--help` flag.
-* Added a validation to the **validate** command that verifies that `unsearchable` field in incident, indicator and generic fields is set to true.
+* Added validation to the **validate** command that verifies the `unsearchable` key in incident, indicator and generic fields is set to true.
 * Removed a validation that DBotRole should be set for automation that requires elevated permissions to the `XSOAR-linter` in the **lint** command.
 * Fixed an issue in **Validate** command where playbooks conditional tasks were mishandeled.
 * Added a validation to prevent contributors from using the `fromlicense` key as a configuration parameter in an integration's YML